requirements.txt
grab==0.6.35
numpy==1.11.2
pytz==2016.7
xmltodict==0.10.2
python_dateutil==2.6.0
mysql_connector_repackaged==0.3.1
scikit_learn==0.18.1
